Extract presence select in its own component
