+++ src/client/components/ParticipantListPage/containers/PresenceFilterContainer.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import { Input } from 'react-bootstrap';
-import { getPresenceLabel } from '../../../components';
+import { PresenceSelector } from '../../../components';
 
 export function getPresenceFilterContainer() {
   const property = 'presence';
@@ -17,17 +16,8 @@ export function getPresenceFilterContainer() {
     }
 
     render() {
-      const presenceLabel = getPresenceLabel(1);
-      const tmpOutCampLabel = getPresenceLabel(2);
-      const outCampLabel = getPresenceLabel(3);
-
       return (
-        <Input type="select" label="Tila" value={ this.props.currentSelection[property] } onChange={ this.onChange }>
-          <option value=""></option>
-          <option value="1">{ presenceLabel }</option>
-          <option value="2">{ tmpOutCampLabel }</option>
-          <option value="3">{ outCampLabel }</option>
-        </Input>
+        <PresenceSelector label="Tila" onChange={ this.onChange } value={ this.props.currentSelection[property] }/>
       );
     }
   }
