+++ src/client/components/index.js
@@ -28,3 +28,4 @@ export { getLogin } from './OfflineLogin';
 export { PropertyTextArea } from './PropertyTextArea';
 export { LoadingButton } from './Util/LoadingButton';
 export { TdWithTitle } from './Util/TdWithTitle';
+export { PresenceSelector } from './Util/PresenceSelector';
